added Repository-service logic. Added Exceptions for handling. Added Debugpy. Fixed tests fixture

+++ app/api/repositories/base.py
@@ -1,78 +1,55 @@
 from typing import Any, Optional
-from fastapi import HTTPException, logger, status
-from sqlalchemy.exc import IntegrityError
+
+from sqlalchemy.sql.elements import BinaryExpression
+from sqlalchemy.engine.result import ScalarResult
 
 from sqlmodel import Session, select
 
 
 class BaseRepository:
+    """
+    Base class to handle all DB interactions
+
+    Attributes:
+        model: The model to be set to which the methods will execute DB operations
+    """
+
     model = None
 
     @classmethod
-    def get_one(
-        cls, filter: Optional[dict] = None, db: Optional[Session] = None
-    ) -> model:
-        try:
-            sql_filter = [getattr(cls.model, k) == v for k, v in filter.items()]
-            response = db.exec(select(cls.model).where(*sql_filter)).first()
-            if not response:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"{cls.model.__name__} Not found",
-                )
-            return response
-        except AttributeError as e:
-            logger.logger.error(e)
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Invalid filter parameter",
-            )
-        except Exception as e:
-            logger.logger.error(e)
-            raise HTTPException(
-                status_code=status.HTTP_418_IM_A_TEAPOT, detail="Unhandled Error"
-            )
-
-    @classmethod
-    def get_all(cls, db: Optional[Session] = None):
-        try:
-            response = db.exec(select(cls.model)).all()
-            if not response:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=f"{cls.model.__name__} Not found",
-                )
-            return response
-
-        except Exception as e:
-            logger.logger.error(e)
-            raise HTTPException(
-                status_code=status.HTTP_418_IM_A_TEAPOT, detail="Unhandled Error"
-            )
+    def get(
+        cls,
+        filter: Optional[list[BinaryExpression]] = None,
+        db: Optional[Session] = None,
+    ) -> ScalarResult:
+        "Returns a `ScalarResult` of the `cls.model` instance with the given `filter`"
+        if not filter:
+            response = db.exec(select(cls.model))
+        else:
+            response = db.exec(select(cls.model).where(*filter))
+        return response
 
     @classmethod
     def create(cls, data: Optional[dict] = None, db: Optional[Session] = None):
-        try:
-            db_model = cls.model(**data)
-            db.add(db_model)
-            db.commit()
-            db.refresh(db_model)
-            return db_model
-        except IntegrityError as e:
-            print(e)
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=f"{cls.model.__name__} already exists",
-            )
-        except Exception as e:
-            print(e)
-            raise HTTPException(
-                status_code=status.HTTP_418_IM_A_TEAPOT, detail="Unhandled Error"
-            )
+        "Creates an instance of the `cls.model` in the DB given the proper `data`"
+        db_model = cls.model(**data)
+        db.add(db_model)
+        db.commit()
+        db.refresh(db_model)
+        return db_model
 
+    # TODO actualizar de aca para abajo y agregarlo al service
     @classmethod
-    def patch(cls, id: int, data: Optional[dict[str, Any]] = None, db: Session = None):
-        db_model = cls.get_one(filter={"id": id}, db=db)
+    def update(
+        cls,
+        filter: Optional[list[BinaryExpression]],
+        data: Optional[dict[str, Any]] = None,
+        db: Session = None,
+    ):
+        "Updates the `cls.model` filtered by `filter` with the proper `data`"
+        db_model = cls.get(filter=filter, db=db).one_or_none()
+        if not db_model:
+            return None
         for model_attr, attr_value in data.items():
             setattr(db_model, model_attr, attr_value)
         db.add(db_model)
@@ -81,14 +58,13 @@ class BaseRepository:
         return db_model
 
     @classmethod
-    def delete(cls, id: int, db: Session = None):
-        try:
-            db_model = cls.get_one(filter={"id": id}, db=db)
-            db.delete(db_model)
-            db.commit()
-            return None
-        except Exception as e:
-            logger.logger.exception(e)
-            raise HTTPException(
-                status_code=status.HTTP_418_IM_A_TEAPOT, detail="Unhandled Error"
-            )
+    def delete(
+        cls, filter: Optional[list[BinaryExpression]], db: Session = None
+    ) -> bool:
+        """Deletes a `cls.model` instance from the DB given the `id`"""
+        db_model = cls.get(filter=filter, db=db).one_or_none()
+        if not db_model:
+            return False
+        db.delete(db_model)
+        db.commit()
+        return True
